Replace set_device with handle_notification API
This will deprecate the set_device API, which was only used by the
RemoteBridge to forward notifications to the MessageBus. set_device is
a bit of a misnomer, because the only use case we had for it was for
forwarding notifications. handle_notification is a more accurate API
and will also only support oneway communication (this reduces the typical
response overhead).