Run lib2to3 with "python3" instead of "python".
Previously, this was almost always using the server/switch-embedded's version of Python installed as
part of the virtualenv. Recently, the build script was changed to not run "activate" so that
stricter failure checking could be added to the setup script. The server version of Ubuntu does not
come with Python2.7 installed, but everyone's localhost and the server should all have pyhton3
installed.