Add constant for external config virtual device
Once we migrate the configuration virtual device off the cloud, we'll
want to host it locally. Our scheme for maintaining consistency is to
have the configuration peripherals merge any notifications about other
config virtual devices in with their current state. To do this, the
peripherals need to listen for those notifications. If they listen to
the normal configuration virtual device, however, there will be a
circular dependency problem. That is, the config peripherals listen to
the notification, merge state, emit their own notification, and then
get notified about their own change since it's the same device. Using a
different name for external config devices will prevent this.