Add a timestamp to the Device struct
The Device struct is one of the few structs that does not have a timestamp. Thus, we'd like to make
it consistent with the other structs. Additionally, this will allow us to store the Device thrift in
PostgreSQL and Redis instead of wrapping the Device in the SubscriptionNotification to provide a
timestamp.

CR: Jeremy