Add device_attributes map to ThirdpartyDeviceInterface
Some thirdparty devices return additional attributes in response to
SSDP discovery requests which are needed/useful to the Coordinators
or Peripherals. Initially we’ll be passing in “hue-bridgeid” for the
Philips Hue integration. But the upcoming change to switch to the new
Sonos API will also make use of this new map.