bluetooth.thrift: Add SLIDE_OVERSHOOT_TO_TURN_OFF property
This configures additional distance past zero required for a slide gesture
to actually turn off the load. The idea is that it should be easy for users
to reach the minimum 'on' value without accidentally turning off the lights,
while still preserving the possibility of using a slide gesture in case of
a confused user who does not know how to otherwise turn the load off.

CR: Steven