Add detailed_discovery_mode to DiscoveryPeripheralInterface.
Variables must be included in the thrift peripheral interface in order for other peripherals to be able to access and set them. In this case we are migrating this variable to the thirdparty_discovery_peripheral. As an intermediate step though, we will need to communicate with the discovery_peripheral via a VariableSubscriptionManager.

https://brillianthome.atlassian.net/browse/CQ-10083

CR: Ashley